+++ tests/model/create_reffed_models_callback_test.py
@@ -1,7 +1,7 @@
 from jsonref import JsonRef
 import pytest
 
-from bravado_core.model import create_reffed_models_callback
+from bravado_core.model import create_reffed_models_callback, MODEL_MARKER
 
 
 @pytest.fixture
@@ -15,7 +15,7 @@ def response_spec():
 @pytest.fixture
 def pet_model_spec():
     return {
-        'x-model': 'Pet',
+        MODEL_MARKER: 'Pet',
         'type': 'object',
         'properties': {
             'name': {
@@ -50,7 +50,7 @@ def test_noop_when_not_jsonref(response_spec, pet_model_spec):
 
 def test_noop_when_not_a_model(response_spec, pet_model_spec):
     models = {}
-    del pet_model_spec['x-model']
+    del pet_model_spec[MODEL_MARKER]
     response_spec['schema'] = build_pet_proxy(pet_model_spec)
     create_reffed_models_callback(models, response_spec, key='schema')
     assert len(models) == 0